+++ backend/controllers/userController.js
@@ -1,0 +1,92 @@
+// controllers/userController.js
+const ConnectionRequest = require("../models/connectionRequest");
+const User = require("../models/user");
+
+const USER_SAFE_DATA = "firstName lastName photoUrl age gender about skills";
+
+// Get all pending connection requests for the logged-in user
+const getPendingRequests = async (req, res) => {
+  try {
+    const loggedInUser = req.user;
+
+    const connectionRequests = await ConnectionRequest.find({
+      toUserId: loggedInUser._id,
+      status: "interested",
+    }).populate("fromUserId", USER_SAFE_DATA);
+
+    res.json({
+      message: "Data fetched successfully",
+      data: connectionRequests,
+    });
+  } catch (err) {
+    res.status(400).send("ERROR: " + err.message);
+  }
+};
+
+// Get all accepted connections for the logged-in user
+const getUserConnections = async (req, res) => {
+  try {
+    const loggedInUser = req.user;
+
+    const connectionRequests = await ConnectionRequest.find({
+      $or: [
+        { toUserId: loggedInUser._id, status: "accepted" },
+        { fromUserId: loggedInUser._id, status: "accepted" },
+      ],
+    })
+      .populate("fromUserId", USER_SAFE_DATA)
+      .populate("toUserId", USER_SAFE_DATA);
+
+    const data = connectionRequests.map((row) =>
+      row.fromUserId._id.toString() === loggedInUser._id.toString()
+        ? row.toUserId
+        : row.fromUserId
+    );
+
+    res.json({ data });
+  } catch (err) {
+    res.status(400).send({ message: err.message });
+  }
+};
+
+// Get user feed (excluding connections and requests)
+const getUserFeed = async (req, res) => {
+  try {
+    const loggedInUser = req.user;
+
+    const page = parseInt(req.query.page) || 1;
+    let limit = parseInt(req.query.limit) || 10;
+    limit = limit > 50 ? 50 : limit;
+    const skip = (page - 1) * limit;
+
+    const connectionRequests = await ConnectionRequest.find({
+      $or: [{ fromUserId: loggedInUser._id }, { toUserId: loggedInUser._id }],
+    }).select("fromUserId toUserId");
+
+    const hideUsersFromFeed = new Set();
+    connectionRequests.forEach((req) => {
+      hideUsersFromFeed.add(req.fromUserId.toString());
+      hideUsersFromFeed.add(req.toUserId.toString());
+    });
+
+    const users = await User.find({
+      $and: [
+        { _id: { $nin: Array.from(hideUsersFromFeed) } },
+        { _id: { $ne: loggedInUser._id } },
+      ],
+    })
+      .select(USER_SAFE_DATA)
+      .skip(skip)
+      .limit(limit);
+
+    res.json({ data: users });
+  } catch (err) {
+    res.status(400).json({ message: err.message });
+  }
+};
+
+module.exports = {
+  getPendingRequests,
+  getUserConnections,
+  getUserFeed,
+};
