+++ backend/routes/request.js
@@ -1,120 +1,11 @@
+// routes/requestRoutes.js
 const express = require("express");
-const requestRouter = express.Router();
-
 const { userAuth } = require("../middlewares/auth");
-const ConnectionRequest = require("../models/connectionRequest");
-const User = require("../models/user");
-const sendPushNotification = require("../config/send_push_notification");
+const { sendConnectionRequest, reviewConnectionRequest } = require("../controllers/requestController");
 
-requestRouter.post(
-  "/request/send/:status/:toUserId",
-  userAuth,
-  async (req, res) => {
-    try {
-      const fromUserId = req.user._id;
-      const toUserId = req.params.toUserId;
-      const status = req.params.status;
+const router = express.Router();
 
-      const allowedStatus = ["ignored", "interested"];
-      // Validation
-      if (!allowedStatus.includes(status)) {
-        return res
-          .status(400)
-          .json({ message: "Invalid status type: " + status });
-      }
+router.post("/request/send/:status/:toUserId", userAuth, sendConnectionRequest);
+router.post("/request/review/:status/:requestId", userAuth, reviewConnectionRequest);
 
-      const toUser = await User.findById(toUserId);
-      if (!toUser) {
-        return res.status(404).json({ message: "User not found!" });
-      }
-
-      const existingConnectionRequest = await ConnectionRequest.findOne({
-        $or: [
-          { fromUserId, toUserId },
-          { fromUserId: toUserId, toUserId: fromUserId },
-        ],
-      });
-      if (existingConnectionRequest) {
-        return res
-          .status(400)
-          .send({ message: "Connection Request Already Exists!!" });
-      }
-
-      const connectionRequest = new ConnectionRequest({
-        fromUserId,
-        toUserId,
-        status,
-      });
-
-      const data = await connectionRequest.save();
-      // send notification if interested
-      if (status == "interested") {
-        await sendPushNotification(
-          toUser.onesignalPlayerId,
-          req.user.firstName + " is " + status + " in you!",
-          process.env.URL_FOR_DEEP_LINK + "/requests"
-        );
-      }
-
-      res.json({
-        message:
-          req.user.firstName + " is " + status + " in " + toUser.firstName,
-        data,
-      });
-    } catch (err) {
-      res.status(400).send("ERROR: " + err.message);
-    }
-  }
-);
-
-requestRouter.post(
-  "/request/review/:status/:requestId",
-  userAuth,
-  async (req, res) => {
-    try {
-      const loggedInUser = req.user;
-      const { status, requestId } = req.params;
-      // validation
-      const allowedStatus = ["accepted", "rejected"];
-      if (!allowedStatus.includes(status)) {
-        return res.status(400).json({ messaage: "Status not allowed!" });
-      }
-
-      const connectionRequest = await ConnectionRequest.findOne({
-        _id: requestId,
-        toUserId: loggedInUser._id,
-        status: "interested",
-      });
-      if (!connectionRequest) {
-        return res
-          .status(404)
-          .json({ message: "Connection request not found" });
-      }
-      // Send notification for user
-      if (status === "accepted") {
-        const fromUser = await User.findOne({
-          _id: connectionRequest.fromUserId,
-        });
-
-        if (!fromUser) {
-          return res.send({ status: 0, message: "Sonthing went wrong" });
-        }
-
-        await sendPushNotification(
-          fromUser.onesignalPlayerId,
-          loggedInUser.firstName + " has " + status + " your request",
-          process.env.URL_FOR_DEEP_LINK + "/connections"
-        );
-      }
-      connectionRequest.status = status;
-
-      const data = await connectionRequest.save();
-
-      res.json({ message: "Connection request " + status, data });
-    } catch (err) {
-      res.status(400).send("ERROR: " + err.message);
-    }
-  }
-);
-
-module.exports = requestRouter;
+module.exports = router;
